(PC-17935)[PRO] chore: split StockFormRow into StockEvent and StockThing
There are some differences and i guess there will be more
2 movements
-> StockFormRow becomes ThingStockFormRow
-> components/StockFormRow/StockFormActions becomes components/StockFormActions
