+++ pro/src/components/StockFormActions/StockFormActions.module.scss
