+++ pro/src/components/StockFormActions/__specs__/StockFormActions.spec.tsx
