--- pro/src/components/StockFormRow/index.ts
@@ -1,1 +1,0 @@
-export { default as StockFormRow } from './StockFormRow'
