+++ pro/src/components/StockThingFormRow/index.ts
@@ -1,0 +1,1 @@
+export { default as StockThingFormRow } from './StockThingFormRow'
